Fixed back buffer race condition.
This change fixes a race condition where the back buffer was
sampled at the same time as it was drawn into. This led to a race
where differently scheduled regions of fragments would sample the
back-buffer texture either before or after that region was
updated, resulting in "boundary" artifacts on both fragment
region edges and polygon boundaries.

The fix involves adding a second render target, and swapping the
buffers between the two render targets in-between frames, such
that the "front" texture is being sampled while the "back"
texture is drawn into.