Implement glowstick preset.
This commit implements a new preset, Glowsticks3d. Currently, this
preset is a misnomer (it uses a 2D ribbon). The implementation includes
a new ribbon GL primitive that uses a shift buffer of segments in a
triangle strip to draw a ribbon. Adding new segments to the buffer
"drags" the ribbon along the screen.