Mitigated program crash on failure to allocate GPU resources.
This change refactors the GL primitive wrappers for textures and render
targets to use absl::StatusOr, so as to enable reporting failure to
allocate the associated GPU resources upstream (instead of
CHECK-failing). This change also implements the associated plumbing to
pass these failre codes all the way up the preset allocation path so as
to enable ignoring the failures at runtime by aborting the preset
allocation.