Updated implementation to have LogLinear interpolation for zoom speed multiplier
This commit changes the behavior of rotary_transporter such that the
zoom_speed is calculated from a function of the signal power which
interpolates back and forth between being a linear and a logarithmic
function of the signal power. The effect of this change is that when
the speed goes as the signal energy, the visual moves at a high rate of
speed with a lot of fine-detail dynamics in the paths traced on the
accumulator raster. When the speed goes as the log of the signal energy,
we retain fine detail, but at a much slower speed (almost static).