Make the ring of squiggles in rotary transporter turn with movement.
This change adds a rotation matrix to the squiggles in rotary
transporter computed from the displacement vector passed to the warp
shader. This makes the ring of squiggles appear to rotate to "face"
along the direction of apparent movement, such that the tunnel and tube
appear to rotate back and forth in the third dimension in front of the
camera.